+++ custom_components/homematicip_local/control_unit.py
@@ -38,6 +38,7 @@ from hahomematic.const import (
 )
 from hahomematic.platforms.device import HmDevice
 from hahomematic.platforms.entity import CallbackEntity
+from hahomematic.support import check_config
 
 from homeassistant.const import CONF_HOST, CONF_PATH, CONF_PORT
 from homeassistant.core import HomeAssistant, callback
@@ -87,6 +88,7 @@ from .support import (
     CLICK_EVENT_SCHEMA,
     DEVICE_AVAILABILITY_EVENT_SCHEMA,
     DEVICE_ERROR_EVENT_SCHEMA,
+    InvalidConfig,
     cleanup_click_event_data,
     get_device_address_at_interface_from_identifiers,
     is_valid_event,
@@ -197,7 +199,7 @@ class BaseControlUnit:
             default_callback_port=self._default_callback_port,
             interface_configs=interface_configs,
             start_direct=self._start_direct,
-        ).create_central()
+        ).create_central(extended_validation=False)
 
 
 class ControlUnit(BaseControlUnit):
@@ -559,6 +561,18 @@ class ControlConfig:
         self.sysvar_scan_interval: Final = data.get(
             CONF_SYSVAR_SCAN_INTERVAL, DEFAULT_SYSVAR_SCAN_INTERVAL
         )
+
+    def check_config(self, extended_validation: bool = True) -> None:
+        """Check config. Throws BaseHomematicException on failure."""
+        if config_failures := check_config(
+            central_name=self.data.get(CONF_INSTANCE_NAME),
+            username=self.data.get(CONF_USERNAME),
+            password=self.data.get(CONF_PASSWORD),
+            storage_folder=get_storage_folder(self.hass),
+            extended_validation=extended_validation,
+        ):
+            failures = ", ".join(config_failures)
+            raise InvalidConfig(failures)
 
     async def create_control_unit(self) -> ControlUnit:
         """Identify the used client."""